requirements.txt
python-telegram-bot==21.5
openai>=1.0.0
python-dotenv
